requirements.txt
streamlit>=1.30.0
groq>=0.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
fpdf2>=2.7.8
